Transfer router from index.js to Application
Allows successful passing of props to other components while also integrating proper routing
