+++ lib/index.js
@@ -1,30 +1,2 @@
-import React from 'react';
-import { render } from 'react-dom';
-import { BrowserRouter, Match, Miss } from 'react-router'
-
-
-import Application from './components/Application';
-import MyMovies from './components/MyMovies';
-import SignIn from './components/SignIn';
-import NotFound from './components/NotFound';
-import Navigation from './components/Navigation';
-import firebase from './firebase';
-
+require ('./components/Application')
 require('./style.scss');
-
-const Root = () => {
-  return(
-    <BrowserRouter>
-      <div>
-        <Match exactly pattern="/" component={Application} />
-        {/* <Match exactly pattern="/" component={MyMovies} /> */}
-        {/* <Match exactly pattern="/" component={App} />
-        <Match exactly pattern="/about" component={About} />
-        <Match exactly pattern="/contact" component={Contact} />*/}
-        <Miss component={NotFound} />
-      </div>
-    </BrowserRouter>
-  )
-}
-
-render(<Root />, document.querySelector("#application"))
